requirements.txt
pandas
pyyaml
psycopg2-binary
lxml
streamlit
Flask
SQLAlchemy
flask_sqlalchemy
matplotlib